Rough Draft of Design
This version should open the form, but the buttons/lists won't work. Just wanted to put this version here to see if it's working for everyone (while I figure out how to get the buttons and lists working).

Also put it in this branch so it doesn't change our original one.